Feat: 네번째 메인페이지 UI 개발
